requirements.txt
fastapi
uvicorn[standard]
python-dotenv
opencv-python-headless
Pillow
pytest
google-generativeai
